+++ examples/cone.example/src/cone/example/__init__.py
@@ -1,9 +1,10 @@
+from cone.app import main_hook
 from cone.app import register_entry
-from cone.app import register_main_hook
 from cone.example.model import ExamplePlugin
 import cone.app
 
 
+@main_hook
 def example_main_hook(config, global_config, local_config):
     """Function which gets called at application startup to initialize
     this plugin.
@@ -22,6 +23,3 @@ def example_main_hook(config, global_config, local_config):
 
     # scan browser package
     config.scan('cone.example.browser')
-
-# register the main hook for this plugin
-register_main_hook(example_main_hook)